refactor: reorganize imports and restructure module exports
