+++ src/__tests__/genres-client.spec.ts
@@ -1,5 +1,5 @@
 import { beforeAll, beforeEach, describe, expect, it } from 'bun:test'
-import { GenresClient } from '../clients'
+import { GenresClient } from '../clients/genres.client'
 import type { Genre, JikanResponse } from '../models'
 
 describe('test Genres Client', () => {
