+++ src/__tests__/top-client.spec.ts
@@ -1,5 +1,5 @@
 import { beforeAll, beforeEach, describe, expect, it } from 'bun:test'
-import { TopClient } from '../clients'
+import { TopClient } from '../clients/top.client'
 import type {
 	Anime,
 	AnimeTopParams,
