--- src/clients/index.ts
@@ -1,10 +1,0 @@
-export * from './anime.client'
-export * from './base.client'
-export * from './characters.client'
-export * from './genres.client'
-export * from './jikan.client'
-export * from './manga.client'
-export * from './schedules.client'
-export * from './seasons.client'
-export * from './top.client'
-export * from './random.client'
