--- src/config/index.ts
@@ -1,2 +1,0 @@
-export * from './logging.config'
-export * from './axios.config'
