--- src/constants/endpoints/index.ts
@@ -1,7 +1,0 @@
-export * from './anime.endpoints'
-export * from './characters.endpoints'
-export * from './genres.endpoints'
-export * from './manga.endpoints'
-export * from './seasons.endpoints'
-export * from './top.endpoints'
-export * from './random.endpoints'
