--- src/constants/index.ts
@@ -1,2 +1,0 @@
-export * from './base.constant'
-export * from './endpoints'
