+++ src/index.ts
@@ -1,4 +1,27 @@
-export * from './clients'
-export * from './config'
-export * from './constants'
+// CLIENTS
+export * from './clients/anime.client'
+export * from './clients/base.client'
+export * from './clients/characters.client'
+export * from './clients/genres.client'
+export * from './clients/jikan.client'
+export * from './clients/manga.client'
+export * from './clients/schedules.client'
+export * from './clients/seasons.client'
+export * from './clients/top.client'
+export * from './clients/random.client'
+
+// CONFIG
+export * from './config/logging.config'
+
+// CONSTANTS
+export * from './constants/endpoints/anime.endpoints'
+export * from './constants/endpoints/characters.endpoints'
+export * from './constants/endpoints/genres.endpoints'
+export * from './constants/endpoints/manga.endpoints'
+export * from './constants/endpoints/seasons.endpoints'
+export * from './constants/endpoints/top.endpoints'
+export * from './constants/endpoints/random.endpoints'
+export * from './constants/base.constant'
+
+// MODELS
 export * from './models'
